Info的Close当因Temporary造成删除，之后的加锁功能自动失效。

+++ system/kernel/src/main/java/indi/sly/system/kernel/objects/prototypes/processors/InfoCloseThenDeleteIfTemporaryResolver.java
@@ -19,17 +19,21 @@ import java.util.List;
 public class InfoCloseThenDeleteIfTemporaryResolver extends AInfoResolver {
     public InfoCloseThenDeleteIfTemporaryResolver() {
         this.close = (info, type, status) -> {
-            if (!status.getIdentifications().isEmpty()) {
-                if (type.isTypeInitializerAttributesExist(TypeInitializerAttributeType.TEMPORARY) && info.getOpened() <= 0) {
-                    List<IdentificationDefinition> identifications = new ArrayList<>(status.getIdentifications());
-                    IdentificationDefinition identification = identifications.remove(identifications.size() - 1);
+            if (!status.getIdentifications().isEmpty()
+                    && type.isTypeInitializerAttributesExist(TypeInitializerAttributeType.TEMPORARY) && info.getOpened() <= 0) {
+                List<IdentificationDefinition> identifications = new ArrayList<>(status.getIdentifications());
+                IdentificationDefinition identification = identifications.remove(identifications.size() - 1);
 
-                    ObjectManager objectManager = this.factoryManager.getManager(ObjectManager.class);
-                    InfoObject parentInfo = objectManager.get(identifications);
+                ObjectManager objectManager = this.factoryManager.getManager(ObjectManager.class);
+                InfoObject parentInfo = objectManager.get(identifications);
 
-                    parentInfo.deleteChild(identification);
-                }
+                parentInfo.deleteChild(identification);
+
+                info = null;
             }
+
+            return info;
+
         };
     }
